requirements.txt
Flask
python-dotenv
psycopg2-binary # For Supabase, if using Postgres
supabase # Supabase Python client
Werkzeug
requests
python-dateutil
pytz

# For testing
pytest
pytest-playwright
